# [libc++] Optimize multi{map,set}::insert(InputIterator, InputIterator)