```
---------------------------------------------------------------------------------------------------------------------------------
Benchmark                                                                                                      old            new
---------------------------------------------------------------------------------------------------------------------------------
std::multimap<int, int>::ctor(iterator, iterator) (unsorted sequence)/0                                    15.0 ns        14.9 ns
std::multimap<int, int>::ctor(iterator, iterator) (unsorted sequence)/32                                    499 ns         453 ns
std::multimap<int, int>::ctor(iterator, iterator) (unsorted sequence)/1024                                48042 ns       45413 ns
std::multimap<int, int>::ctor(iterator, iterator) (unsorted sequence)/8192                               720794 ns      699102 ns
std::multimap<int, int>::ctor(iterator, iterator) (sorted sequence)/0                                      14.9 ns        15.0 ns
std::multimap<int, int>::ctor(iterator, iterator) (sorted sequence)/32                                      469 ns         312 ns
std::multimap<int, int>::ctor(iterator, iterator) (sorted sequence)/1024                                  24870 ns       12535 ns
std::multimap<int, int>::ctor(iterator, iterator) (sorted sequence)/8192                                 320474 ns       98615 ns
std::multimap<int, int>::insert(iterator, iterator) (all new keys)/0                                        455 ns         462 ns
std::multimap<int, int>::insert(iterator, iterator) (all new keys)/32                                      1069 ns        1024 ns
std::multimap<int, int>::insert(iterator, iterator) (all new keys)/1024                                   29448 ns       29650 ns
std::multimap<int, int>::insert(iterator, iterator) (all new keys)/8192                                  316087 ns      307652 ns
std::multimap<int, int>::insert(iterator, iterator) (half new keys)/0                                       456 ns         459 ns
std::multimap<int, int>::insert(iterator, iterator) (half new keys)/32                                      982 ns         926 ns
std::multimap<int, int>::insert(iterator, iterator) (half new keys)/1024                                  40963 ns       35205 ns
std::multimap<int, int>::insert(iterator, iterator) (half new keys)/8192                                 470584 ns      411297 ns
std::multimap<int, int>::insert(iterator, iterator) (product_iterator from same type)/0                     455 ns         455 ns
std::multimap<int, int>::insert(iterator, iterator) (product_iterator from same type)/32                    981 ns         908 ns
std::multimap<int, int>::insert(iterator, iterator) (product_iterator from same type)/1024                29573 ns       18495 ns
std::multimap<int, int>::insert(iterator, iterator) (product_iterator from same type)/8192               373138 ns      145490 ns
std::multimap<int, int>::insert(iterator, iterator) (product_iterator from zip_view)/0                      456 ns         457 ns
std::multimap<int, int>::insert(iterator, iterator) (product_iterator from zip_view)/32                    1001 ns         846 ns
std::multimap<int, int>::insert(iterator, iterator) (product_iterator from zip_view)/1024                 28886 ns       16380 ns
std::multimap<int, int>::insert(iterator, iterator) (product_iterator from zip_view)/8192                296621 ns      129634 ns
std::multimap<int, int>::erase(iterator, iterator) (erase half the container)/0                             461 ns         459 ns
std::multimap<int, int>::erase(iterator, iterator) (erase half the container)/32                            712 ns         721 ns
std::multimap<int, int>::erase(iterator, iterator) (erase half the container)/1024                         8683 ns        9032 ns
std::multimap<int, int>::erase(iterator, iterator) (erase half the container)/8192                        67316 ns       70256 ns
std::multimap<std::string, int>::ctor(iterator, iterator) (unsorted sequence)/0                            15.0 ns        17.0 ns
std::multimap<std::string, int>::ctor(iterator, iterator) (unsorted sequence)/32                           3133 ns        3079 ns
std::multimap<std::string, int>::ctor(iterator, iterator) (unsorted sequence)/1024                       189504 ns      190940 ns
std::multimap<std::string, int>::ctor(iterator, iterator) (unsorted sequence)/8192                      2578482 ns     2534645 ns
std::multimap<std::string, int>::ctor(iterator, iterator) (sorted sequence)/0                              15.2 ns        16.9 ns
std::multimap<std::string, int>::ctor(iterator, iterator) (sorted sequence)/32                             2690 ns        2189 ns
std::multimap<std::string, int>::ctor(iterator, iterator) (sorted sequence)/1024                          98889 ns       71348 ns
std::multimap<std::string, int>::ctor(iterator, iterator) (sorted sequence)/8192                        1274108 ns     1159230 ns
std::multimap<std::string, int>::insert(iterator, iterator) (all new keys)/0                                455 ns         456 ns
std::multimap<std::string, int>::insert(iterator, iterator) (all new keys)/32                              2007 ns        2076 ns
std::multimap<std::string, int>::insert(iterator, iterator) (all new keys)/1024                          160027 ns      156791 ns
std::multimap<std::string, int>::insert(iterator, iterator) (all new keys)/8192                         1540491 ns     1470501 ns
std::multimap<std::string, int>::insert(iterator, iterator) (half new keys)/0                               455 ns         457 ns
std::multimap<std::string, int>::insert(iterator, iterator) (half new keys)/32                             2003 ns        2063 ns
std::multimap<std::string, int>::insert(iterator, iterator) (half new keys)/1024                         163551 ns      167734 ns
std::multimap<std::string, int>::insert(iterator, iterator) (half new keys)/8192                        1592041 ns     1519982 ns
std::multimap<std::string, int>::insert(iterator, iterator) (product_iterator from same type)/0             452 ns         451 ns
std::multimap<std::string, int>::insert(iterator, iterator) (product_iterator from same type)/32           1647 ns        1553 ns
std::multimap<std::string, int>::insert(iterator, iterator) (product_iterator from same type)/1024       115961 ns       82969 ns
std::multimap<std::string, int>::insert(iterator, iterator) (product_iterator from same type)/8192       805502 ns      809357 ns
std::multimap<std::string, int>::insert(iterator, iterator) (product_iterator from zip_view)/0              456 ns         461 ns
std::multimap<std::string, int>::insert(iterator, iterator) (product_iterator from zip_view)/32            1599 ns        1569 ns
std::multimap<std::string, int>::insert(iterator, iterator) (product_iterator from zip_view)/1024        105381 ns       83137 ns
std::multimap<std::string, int>::insert(iterator, iterator) (product_iterator from zip_view)/8192        769359 ns      627449 ns
std::multimap<std::string, int>::erase(iterator, iterator) (erase half the container)/0                     454 ns         459 ns
std::multimap<std::string, int>::erase(iterator, iterator) (erase half the container)/32                    831 ns         826 ns
std::multimap<std::string, int>::erase(iterator, iterator) (erase half the container)/1024                22997 ns       22864 ns
std::multimap<std::string, int>::erase(iterator, iterator) (erase half the container)/8192               138008 ns      137066 ns
std::multiset<int>::ctor(iterator, iterator) (unsorted sequence)/0                                         15.0 ns        14.7 ns
std::multiset<int>::ctor(iterator, iterator) (unsorted sequence)/32                                         529 ns         425 ns
std::multiset<int>::ctor(iterator, iterator) (unsorted sequence)/1024                                     49784 ns       42171 ns
std::multiset<int>::ctor(iterator, iterator) (unsorted sequence)/8192                                    725039 ns      699864 ns
std::multiset<int>::ctor(iterator, iterator) (sorted sequence)/0                                           14.9 ns        14.8 ns
std::multiset<int>::ctor(iterator, iterator) (sorted sequence)/32                                           513 ns         354 ns
std::multiset<int>::ctor(iterator, iterator) (sorted sequence)/1024                                       25201 ns       12103 ns
std::multiset<int>::ctor(iterator, iterator) (sorted sequence)/8192                                      320001 ns       96768 ns
std::multiset<int>::insert(iterator, iterator) (all new keys)/0                                             452 ns         458 ns
std::multiset<int>::insert(iterator, iterator) (all new keys)/32                                           1048 ns        1026 ns
std::multiset<int>::insert(iterator, iterator) (all new keys)/1024                                        29217 ns       29509 ns
std::multiset<int>::insert(iterator, iterator) (all new keys)/8192                                       310906 ns      309259 ns
std::multiset<int>::insert(iterator, iterator) (half new keys)/0                                            462 ns         454 ns
std::multiset<int>::insert(iterator, iterator) (half new keys)/32                                           988 ns         908 ns
std::multiset<int>::insert(iterator, iterator) (half new keys)/1024                                       41545 ns       32431 ns
std::multiset<int>::insert(iterator, iterator) (half new keys)/8192                                      470042 ns      409618 ns
std::multiset<int>::erase(iterator, iterator) (erase half the container)/0                                  453 ns         453 ns
std::multiset<int>::erase(iterator, iterator) (erase half the container)/32                                 725 ns         717 ns
std::multiset<int>::erase(iterator, iterator) (erase half the container)/1024                              8734 ns        8848 ns
std::multiset<int>::erase(iterator, iterator) (erase half the container)/8192                             70278 ns       68465 ns
std::multiset<std::string>::ctor(iterator, iterator) (unsorted sequence)/0                                 15.4 ns        16.5 ns
std::multiset<std::string>::ctor(iterator, iterator) (unsorted sequence)/32                                3060 ns        1894 ns
std::multiset<std::string>::ctor(iterator, iterator) (unsorted sequence)/1024                            185144 ns      180302 ns
std::multiset<std::string>::ctor(iterator, iterator) (unsorted sequence)/8192                           2780124 ns     1883800 ns
std::multiset<std::string>::ctor(iterator, iterator) (sorted sequence)/0                                   15.6 ns        16.8 ns
std::multiset<std::string>::ctor(iterator, iterator) (sorted sequence)/32                                  2694 ns        2395 ns
std::multiset<std::string>::ctor(iterator, iterator) (sorted sequence)/1024                               85031 ns       80462 ns
std::multiset<std::string>::ctor(iterator, iterator) (sorted sequence)/8192                              854175 ns     1122345 ns
std::multiset<std::string>::insert(iterator, iterator) (all new keys)/0                                     459 ns         457 ns
std::multiset<std::string>::insert(iterator, iterator) (all new keys)/32                                   2120 ns        1999 ns
std::multiset<std::string>::insert(iterator, iterator) (all new keys)/1024                               161187 ns      141931 ns
std::multiset<std::string>::insert(iterator, iterator) (all new keys)/8192                              1416852 ns     1343619 ns
std::multiset<std::string>::insert(iterator, iterator) (half new keys)/0                                    459 ns         455 ns
std::multiset<std::string>::insert(iterator, iterator) (half new keys)/32                                  2022 ns        1927 ns
std::multiset<std::string>::insert(iterator, iterator) (half new keys)/1024                              169273 ns      152779 ns
std::multiset<std::string>::insert(iterator, iterator) (half new keys)/8192                             1640003 ns     1489332 ns
std::multiset<std::string>::erase(iterator, iterator) (erase half the container)/0                          453 ns         452 ns
std::multiset<std::string>::erase(iterator, iterator) (erase half the container)/32                         827 ns         806 ns
std::multiset<std::string>::erase(iterator, iterator) (erase half the container)/1024                     20004 ns       20002 ns
std::multiset<std::string>::erase(iterator, iterator) (erase half the container)/8192                    135073 ns      134105 ns
```